requirements.txt
plotly == 5.24.1
streamlit_plotly_events == 0.0.6
openpyxl
statsmodels
